dsk-5315 correçao no tipo do campo desconto_id

+++ functions/001-fn_venda_item.sql
@@ -52,7 +52,7 @@ returns @tbl table
   perfil_impressao_id int,
   classe_rodizio_id int,
   transferido bit,
-  desconto_id smallint,
+  desconto_id int,
   desconto_estrategia varchar(100),
   vl_desconto numeric(15, 2),
   slot_indice smallint,
